+++ ios/Classes/FlutterNavigationBarPlugin.h
@@ -2,3 +2,6 @@
 
 @interface FlutterNavigationBarPlugin : NSObject<FlutterPlugin>
 @end
+
+@interface FirstStreamHandler : NSObject<FlutterStreamHandler>
+@end