+++ CHANGELOG.md
@@ -1,3 +1,16 @@
 ## 0.0.1
+Call setNavigationBarTransparent() and this plugin will set activity.getWindow().setFlags(WindowManager.LayoutParams.FLAG_LAYOUT_IN_OVERSCAN, WindowManager.LayoutParams.FLAG_LAYOUT_IN_OVERSCAN)
 
-* Call setNavigationBarTransparent() and this plugin will set activity.getWindow().setFlags(WindowManager.LayoutParams.FLAG_LAYOUT_IN_OVERSCAN, WindowManager.LayoutParams.FLAG_LAYOUT_IN_OVERSCAN)
+## 0.0.2
+Added FLAG_LAYOUT_NO_LIMITS.
+
+## 0.0.3
+Minor change.
+
+## 0.0.4
+Flags are set on init.
+
+## 0.0.5
+Converted to Kotlin and Swift.
+Solved the problem with a transparent Navigation Bar (full screen) and still allowing for the use of safe areas that work with input.
+Added support for equal width sides in landscape.